sched: EAS: update trg_cpu to backup_cpu if no energy saving for target_cpu
If no energy saving for target_cpu in the calculation of energy_diff(),
backup_cpu will be set as the new dst_cpu for the next calculation. At this
point, we also need update the new trg_cpu as backup_cpu to make sure the
subsequent calculation of energy_diff() is correct.

Signed-off-by: Ke Wang <ke.wang@spreadtrum.com>
Signed-off-by: Mohammad Hasan Keramat J <ikeramat80@gmail.com>
Change-Id: I3e7ea6878d6a392edc4c5711ece2e7c9378cb55f